requirements.txt
asgiref==3.11.0
Django==5.2.8
djangorestframework==3.15.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.10.1
sqlparse==0.5.3
gunicorn==20.1.0
python-dotenv==1.0.0
